+++ app/src/main/java/dev/pandesal/sbp/presentation/AppNavigation.kt
@@ -21,6 +21,7 @@ import dev.pandesal.sbp.presentation.transactions.newtransaction.NewRecurringTra
 import dev.pandesal.sbp.presentation.transactions.details.TransactionDetailsScreen
 import dev.pandesal.sbp.presentation.settings.SettingsScreen
 import dev.pandesal.sbp.presentation.notifications.NotificationCenterScreen
+import dev.pandesal.sbp.presentation.goals.NewGoalScreen
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -45,6 +46,8 @@ sealed class NavigationDestination() {
     data object Accounts : NavigationDestination()
     @Serializable
     data object NewAccount : NavigationDestination()
+    @Serializable
+    data object NewGoal : NavigationDestination()
     @Serializable
     data object NewRecurringTransaction : NavigationDestination()
     @Serializable
@@ -79,6 +82,12 @@ fun AppNavigation(navController: NavHostController) {
             ) {
                 NewAccountScreen()
             }
+            dialog<NavigationDestination.NewGoal>(
+                dialogProperties = DialogProperties(usePlatformDefaultWidth = false),
+            ) {
+                NewGoalScreen()
+            }
+
 
             composable<NavigationDestination.Transactions> {
                 TransactionsScreen()